@@ -3,7 +3,7 @@
 	"description": "The Kirby 3 core",
 	"license": "proprietary",
 	"type": "kirby-cms",
-	"version": "3.9.7",
+	"version": "3.10.1",
 	"keywords": [
 		"kirby",
 		"cms",
@@ -24,7 +24,7 @@
 		"source": "https://github.com/getkirby/kirby"
 	},
 	"require": {
-		"php": ">=8.0.0 <8.3.0",
+		"php": "~8.1.0 || ~8.2.0 || ~8.3.0",
 		"ext-SimpleXML": "*",
 		"ext-ctype": "*",
 		"ext-curl": "*",
@@ -36,16 +36,16 @@
 		"ext-libxml": "*",
 		"ext-mbstring": "*",
 		"ext-openssl": "*",
-		"claviska/simpleimage": "4.0.6",
+		"claviska/simpleimage": "4.2.0",
 		"composer/semver": "3.4.0",
-		"filp/whoops": "2.15.3",
+		"filp/whoops": "2.15.4",
 		"getkirby/composer-installer": "^1.2.1",
-		"laminas/laminas-escaper": "2.12.0",
+		"laminas/laminas-escaper": "2.13.0",
 		"michelf/php-smartypants": "1.8.1",
-		"phpmailer/phpmailer": "6.8.1",
-		"symfony/polyfill-intl-idn": "1.28.0",
-		"symfony/polyfill-mbstring": "1.28.0",
-		"symfony/yaml": "5.4.23"
+		"phpmailer/phpmailer": "6.9.1",
+		"symfony/polyfill-intl-idn": "1.30.0",
+		"symfony/polyfill-mbstring": "1.30.0",
+		"symfony/yaml": "6.4.8"
 	},
 	"replace": {
 		"symfony/polyfill-php72": "*"
@@ -57,6 +57,7 @@
 		"ext-fileinfo": "Improved mime type detection for files",
 		"ext-intl": "Improved i18n number formatting",
 		"ext-memcached": "Support for the Memcached cache driver",
+		"ext-sodium": "Support for the crypto class and more robust session handling",
 		"ext-zip": "Support for ZIP archive file functions",
 		"ext-zlib": "Sanitization and validation for svgz files"
 	},
@@ -78,7 +79,7 @@
 		},
 		"optimize-autoloader": true,
 		"platform": {
-			"php": "8.0.0"
+			"php": "8.1.0"
 		},
 		"platform-check": false
 	},
