@@ -16,18 +16,15 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
-        "symfony/deprecation-contracts": "^2.1|^3",
+        "php": ">=8.1",
+        "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/polyfill-ctype": "^1.8"
     },
     "require-dev": {
-        "symfony/console": "^5.3|^6.0"
+        "symfony/console": "^5.4|^6.0|^7.0"
     },
     "conflict": {
-        "symfony/console": "<5.3"
-    },
-    "suggest": {
-        "symfony/console": "For validating YAML files using the lint command"
+        "symfony/console": "<5.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Yaml\\": "" },
