@@ -5,8 +5,8 @@
  * stop at older or too recent versions
  */
 if (
-	version_compare(PHP_VERSION, '8.0.0', '>=') === false ||
-	version_compare(PHP_VERSION, '8.3.0', '<')  === false
+	version_compare(PHP_VERSION, '8.1.0', '>=') === false ||
+	version_compare(PHP_VERSION, '8.4.0', '<')  === false
 ) {
 	die(include __DIR__ . '/views/php.php');
 }
