@@ -417,8 +417,8 @@ class System
 	public function php(): bool
 	{
 		return
-			version_compare(PHP_VERSION, '8.0.0', '>=') === true &&
-			version_compare(PHP_VERSION, '8.3.0', '<')  === true;
+			version_compare(PHP_VERSION, '8.1.0', '>=') === true &&
+			version_compare(PHP_VERSION, '8.4.0', '<')  === true;
 	}
 
 	/**
